@@ -4,25 +4,25 @@
     "homepage": "https://an1mus.github.io/merntodo",
     "private": true,
     "dependencies": {
-        "@testing-library/jest-dom": "^4.2.4",
-        "@testing-library/react": "^9.3.2",
-        "@testing-library/user-event": "^7.1.2",
-        "@types/history": "^4.7.4",
-        "@types/jest": "^24.0.0",
-        "@types/node": "^12.0.0",
-        "@types/react": "^16.9.0",
-        "@types/react-dom": "^16.9.0",
-        "@types/react-redux": "^7.1.5",
-        "@types/styled-components": "^4.4.2",
-        "history": "^4.10.1",
-        "react": "^16.12.0",
-        "react-dom": "^16.12.0",
-        "react-redux": "^7.1.3",
-        "react-router-dom": "^5.1.2",
-        "react-scripts": "3.3.0",
-        "redux-thunk": "^2.3.0",
-        "styled-components": "^5.0.0",
-        "typescript": "~3.7.2"
+        "@testing-library/jest-dom": "^5.16.5",
+        "@testing-library/react": "^13.4.0",
+        "@testing-library/user-event": "^14.4.3",
+        "@types/history": "^4.7.11",
+        "@types/jest": "^29.2.5",
+        "@types/node": "^18.11.18",
+        "@types/react": "^18.0.26",
+        "@types/react-dom": "^18.0.10",
+        "@types/react-redux": "^7.1.25",
+        "@types/styled-components": "^5.1.26",
+        "history": "^5.3.0",
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
+        "react-redux": "^8.0.5",
+        "react-router-dom": "^6.6.1",
+        "react-scripts": "5.0.1",
+        "redux-thunk": "^2.4.2",
+        "styled-components": "^5.3.6",
+        "typescript": "~4.9.4"
     },
     "scripts": {
         "start": "react-scripts start",
@@ -48,6 +48,6 @@
         ]
     },
     "devDependencies": {
-        "gh-pages": "^2.2.0"
+        "gh-pages": "^4.0.0"
     }
 }
